move AuthLayout to ui folder
